requirements.txt
transformers==4.29.2
torch==2.0.1
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.3.0
flask==2.3.3
beautifulsoup4==4.12.2
requests==2.31.0
vaderSentiment==3.3.2
nltk==3.8.1
tqdm==4.65.0
python-dotenv==1.0.0
pyarrow==12.0.1
seaborn==0.12.2
pytest==7.4.0
tweepy==4.14.0
datasets==2.13.1
huggingface-hub==0.16.4
accelerate==0.21.0
evaluate==0.4.0 